Restructured module: added utils/ and docs/ dirs
