+++ docs/CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - Unreleased
+## [1.2.0-dev] - 2023-08-11
+
+### Breaking Changes
+
+- Restructured entire module: created mcswitch/utils/, mcswitch/server.utils, mcswitch/docs folders to contain source code
+- Changed most instances of importing helpers.py: imports are much more streamlined
 
 ### Added
 
 - !define command to get the Meriam-Webster definition of a word
+- !help command description: !define
 
 ## [1.1.6] - 2023-08-11
