+++ main.py
@@ -1,24 +1,25 @@
-import helpers
+from utils.helpers import bot, TOKEN, CHANNEL, COGS
 
-bot = helpers.bot
-TOKEN = helpers.TOKEN
-CHANNEL = helpers.CHANNEL
-COGS = helpers.COGS
+bot = bot
+token = TOKEN
+channel = CHANNEL
+cogs = COGS
 
-bot.remove_command('help')
+bot.remove_command("help")
+
 
 @bot.event
 async def setup_hook() -> None:
     """Loads cogs."""
     print(f"Loaded the following cogs: ")
-    for cog in COGS:
+    for cog in cogs:
         print(cog)
         await bot.load_extension(cog)
 
 
 def main():
     """Start the bot."""
-    bot.run(TOKEN)
+    bot.run(token)
 
 
 main()
