+++ server.utils/ip.sh
@@ -1,0 +1,4 @@
+#!/bin/zsh
+
+# Get public IP
+curl https://ipinfo.io/ip >> ip.txt
